Trimming library
