+++ source/Dgraph-dotnet/Transactions/TransactionFactory.cs
@@ -4,18 +4,12 @@ namespace DgraphDotNet.Transactions
 {
     internal interface ITransactionFactory {
         ITransaction NewTransaction(DgraphClient client);
-
-        ITransactionWithMutations NewTransaction(DgraphMutationsClient client);
     }
 
     internal class TransactionFactory : ITransactionFactory
     {
         public ITransaction NewTransaction(DgraphClient client) {
             return new Transaction(client);
-        }
-
-        public ITransactionWithMutations NewTransaction(DgraphMutationsClient client) {
-            return new TransactionWithMutations(client);
         }
     }
 }
